+++ sdk/solana/package.json
@@ -1,6 +1,7 @@
 {
   "name": "@xlabs-xyz/solana-arbitrary-token-transfers",
   "version": "0.0.1-alpha1",
+  "license": "ISC",
   "main": "tbrv3/index.ts",
   "scripts": {
     "lint:fix": "prettier */*.js \"*/**/*{.js,.ts}\" -w",
@@ -9,8 +10,10 @@
   "dependencies": {
     "@coral-xyz/anchor": "^0.30.1",
     "@solana/web3.js": "^1.95.3",
-    "@wormhole-foundation/sdk-base": "^0.10.0",
-    "solana-price-oracle-sdk": "ssh://git@github.com/XLabs/relayer-infra-contracts#head=solana-oracle"
+    "@wormhole-foundation/sdk-base": "^0.10.4",
+    "@wormhole-foundation/sdk-definitions": "^0.10.4",
+    "@wormhole-foundation/sdk-solana-tokenbridge": "^0.10.4",
+    "solana-price-oracle-sdk": "ssh://git@github.com/XLabs/relayer-infra-contracts#head=solana-program"
   },
   "devDependencies": {
     "@types/bn.js": "^5.1.0",
