+++ solana/programs/token-bridge-relayer/Cargo.toml
@@ -24,8 +24,7 @@ idl-build = [
 anchor-lang = { version = "0.30.1", git = "https://github.com/coral-xyz/anchor/", rev = "06527e57c3e59683c36eb0a5c69ee669100b42e5", features = ["init-if-needed"] }
 anchor-spl = { version = "0.30.1", git = "https://github.com/coral-xyz/anchor/", rev = "06527e57c3e59683c36eb0a5c69ee669100b42e5" }
 
-solana-price-oracle = { git = "ssh://git@github.com/XLabs/relayer-infra-contracts", branch = "solana-oracle" }
-#solana-quoter-program = { path = "../../../../relayer-infra-contracts/src/solana/programs/quoter" } # → anchor-lang 06527e57c3e59683c36eb0a5c69ee669100b42e5
+solana-price-oracle = { git = "ssh://git@github.com/XLabs/relayer-infra-contracts", branch = "solana-program" }
 
 wormhole-anchor-sdk = { git = "https://github.com/wormhole-foundation/wormhole-scaffolding.git", rev = "1acd10e3cbeea0df06976b606bd24da2e8e849e1", features = ["token-bridge"] }
 #wormhole-anchor-sdk = { path = "../../../../wormhole-scaffolding/solana/modules/wormhole-anchor-sdk/", features = ["token-bridge"] }
