+++ solana/programs/token-bridge-relayer/src/processor/initialize.rs
@@ -1,4 +1,4 @@
-use crate::state::TbrConfigAccount;
+use crate::state::TbrConfigState;
 use anchor_lang::prelude::*;
 
 #[derive(Accounts)]
@@ -12,25 +12,22 @@ pub struct Initialize<'info> {
     #[account(
         init,
         payer = owner,
-        space = 8 + TbrConfigAccount::INIT_SPACE,
-        seeds = [TbrConfigAccount::SEED_PREFIX],
+        space = 8 + TbrConfigState::INIT_SPACE,
+        seeds = [TbrConfigState::SEED_PREFIX],
         bump
     )]
-    pub tbr_config: Account<'info, TbrConfigAccount>,
+    pub tbr_config: Account<'info, TbrConfigState>,
 
     pub system_program: Program<'info, System>,
 }
 
-pub fn initialize(ctx: Context<Initialize>, quoter_program: Pubkey) -> Result<()> {
+pub fn initialize(ctx: Context<Initialize>) -> Result<()> {
     let owner = ctx.accounts.owner.key();
     let tbr_config = &mut ctx.accounts.tbr_config;
 
     tbr_config.owner = owner;
     tbr_config.admin = owner;
     tbr_config.pending_owner = None;
-    tbr_config.quoter_program_address = quoter_program;
-    tbr_config.paused_outbound_transfers = true;
-    tbr_config.relayer_fee = 0;
     tbr_config.fee_recipient = owner;
     tbr_config.evm_transaction_size = 0;
