+++ solana/programs/token-bridge-relayer/src/processor/peers.rs
@@ -1,6 +1,6 @@
 use crate::{
     error::TokenBridgeRelayerError,
-    state::{ChainConfigAccount, PeerAccount, TbrConfigAccount},
+    state::{ChainConfigState, PeerState, TbrConfigState},
 };
 use anchor_lang::prelude::*;
 
@@ -10,44 +10,61 @@ use anchor_lang::prelude::*;
     peer_address: [u8; 32],
 )]
 pub struct RegisterPeer<'info> {
-    /// Owner or  of the program as set in the [`TbrConfig`] account.
+    /// Owner or admin of the program as set in the [`TbrConfig`] account.
     #[account(mut)]
     pub signer: Signer<'info>,
 
     /// Owner Config account. This program requires that the `signer` specified
     /// in the context equals an authorized pubkey specified in this account.
     #[account(
-        seeds = [TbrConfigAccount::SEED_PREFIX],
+        seeds = [TbrConfigState::SEED_PREFIX],
         bump
     )]
-    pub tbr_config: Account<'info, TbrConfigAccount>,
+    pub tbr_config: Account<'info, TbrConfigState>,
 
     #[account(
         init,
         payer = signer,
-        space = 8 + PeerAccount::INIT_SPACE,
+        space = 8 + PeerState::INIT_SPACE,
         seeds = [
-            PeerAccount::SEED_PREFIX,
+            PeerState::SEED_PREFIX,
             chain_id.to_be_bytes().as_ref(),
             peer_address.as_ref(),
         ],
         bump
     )]
-    pub peer_account: Account<'info, PeerAccount>,
+    pub peer: Account<'info, PeerState>,
 
     #[account(
         init_if_needed,
         payer = signer,
-        space = 8 + ChainConfigAccount::INIT_SPACE,
+        space = 8 + ChainConfigState::INIT_SPACE,
         seeds = [
-            ChainConfigAccount::SEED_PREFIX,
+            ChainConfigState::SEED_PREFIX,
             chain_id.to_be_bytes().as_ref(),
         ],
         bump
     )]
-    pub chain_config: Account<'info, ChainConfigAccount>,
+    pub chain_config: Account<'info, ChainConfigState>,
 
     pub system_program: Program<'info, System>,
+}
+
+pub fn register_peer(ctx: Context<RegisterPeer>, peer_address: [u8; 32]) -> Result<()> {
+    require!(
+        ctx.accounts
+            .tbr_config
+            .is_owner_or_admin(ctx.accounts.signer.key),
+        TokenBridgeRelayerError::OwnerOrAdminOnly
+    );
+
+    // If it is the first peer for this chain, make it canonical:
+    if ctx.accounts.chain_config.is_uninitialized() {
+        let canonical_peer = &mut ctx.accounts.chain_config;
+        canonical_peer.canonical_peer = peer_address;
+    }
+
+    Ok(())
 }
 
 #[derive(Accounts)]
@@ -55,8 +72,8 @@ pub struct RegisterPeer<'info> {
     chain_id: u16,
     peer_address: [u8; 32],
 )]
-pub struct UpdateCanonical<'info> {
-    /// Owner or  of the program as set in the [`TbrConfig`] account.
+pub struct UpdateCanonicalPeer<'info> {
+    /// Owner of the program as set in the [`TbrConfig`] account.
     #[account(mut)]
     pub owner: Signer<'info>,
 
@@ -64,62 +81,47 @@ pub struct UpdateCanonical<'info> {
     /// in the context equals the `owner` pubkey specified in this account.
     #[account(
         has_one = owner @ TokenBridgeRelayerError::OwnerOnly,
-        seeds = [TbrConfigAccount::SEED_PREFIX],
+        seeds = [TbrConfigState::SEED_PREFIX],
         bump
     )]
-    pub tbr_config: Account<'info, TbrConfigAccount>,
+    pub tbr_config: Account<'info, TbrConfigState>,
 
     #[account(
         seeds = [
-            PeerAccount::SEED_PREFIX,
+            PeerState::SEED_PREFIX,
             chain_id.to_be_bytes().as_ref(),
             peer_address.as_ref(),
         ],
         bump
     )]
-    pub peer_account: Account<'info, PeerAccount>,
+    pub peer: Account<'info, PeerState>,
 
     #[account(
         seeds = [
-            ChainConfigAccount::SEED_PREFIX,
+            ChainConfigState::SEED_PREFIX,
             chain_id.to_be_bytes().as_ref(),
         ],
         bump
     )]
-    pub chain_config: Account<'info, ChainConfigAccount>,
+    pub chain_config: Account<'info, ChainConfigState>,
 
     pub system_program: Program<'info, System>,
 }
 
-pub fn register_peer(ctx: Context<RegisterPeer>, peer_address: [u8; 32]) -> Result<()> {
-    // Verify the authority (owner or admin):
-    require!(
-        ctx.accounts
-            .tbr_config
-            .is_authorized(ctx.accounts.signer.key),
-        TokenBridgeRelayerError::AuthorizedOnly
-    );
-
-    // If it is the first peer for this chain, make it canonical:
-    if ctx.accounts.chain_config.is_uninitialized() {
-        let canonical_peer = &mut ctx.accounts.chain_config;
-        canonical_peer.canonical_peer_address = peer_address;
-    }
-
-    Ok(())
-}
-
-pub fn update_canonical(ctx: Context<UpdateCanonical>, peer_address: [u8; 32]) -> Result<()> {
+pub fn update_canonical_peer(
+    ctx: Context<UpdateCanonicalPeer>,
+    peer_address: [u8; 32],
+) -> Result<()> {
     // Owner authorization verified in the context.
 
     // Verify it is a different peer being set as canonical:
-    if ctx.accounts.chain_config.canonical_peer_address == peer_address {
+    if ctx.accounts.chain_config.canonical_peer == peer_address {
         Err(TokenBridgeRelayerError::AlreadyTheCanonicalPeer)?
     }
 
     // Update the field:
     let canonical_peer = &mut ctx.accounts.chain_config;
-    canonical_peer.canonical_peer_address = peer_address;
+    canonical_peer.canonical_peer = peer_address;
 
     Ok(())
 }
