+++ solana/programs/token-bridge-relayer/src/state/chain.rs
@@ -1,28 +1,42 @@
+use crate::error::TokenBridgeRelayerError;
 use anchor_lang::prelude::*;
 
 /// A peer chain. Nothing is stored in it for now.
 #[account]
 #[derive(InitSpace)]
-pub struct PeerAccount {}
+pub struct PeerState {}
 
-/// The config for every chain.
+/// The config for a single chain.
 #[account]
 #[derive(InitSpace)]
-pub struct ChainConfigAccount {
-    pub canonical_peer_address: [u8; 32],
+pub struct ChainConfigState {
+    pub canonical_peer: [u8; 32],
     pub max_gas_dropoff: u64,
+    pub paused_outbound_transfers: bool,
+    /// The fee for the relayer, in μUSD.
+    pub relayer_fee: u64,
 }
 
-impl PeerAccount {
+impl PeerState {
     /// AKA `b"peer"`.
     pub const SEED_PREFIX: &'static [u8; 4] = b"peer";
 }
 
-impl ChainConfigAccount {
+impl ChainConfigState {
     pub fn is_uninitialized(&self) -> bool {
-        self.canonical_peer_address == [0; 32]
+        self.canonical_peer == [0; 32]
     }
 
-    /// AKA `b"canonical"`.
-    pub const SEED_PREFIX: &'static [u8; 9] = b"canonical";
+    /// Returns an error if the transfers are paused for this chain.
+    pub fn transfer_allowed(&self) -> Result<()> {
+        require!(
+            false == self.paused_outbound_transfers,
+            TokenBridgeRelayerError::PausedTransfers
+        );
+
+        Ok(())
+    }
+
+    /// AKA `b"chainconfig"`.
+    pub const SEED_PREFIX: &'static [u8; 11] = b"chainconfig";
 }
