+++ solana/programs/token-bridge-relayer/src/state/config.rs
@@ -1,9 +1,10 @@
 use anchor_lang::prelude::*;
+use wormhole_anchor_sdk::token_bridge;
 
 /// The program's main account.
 #[account]
 #[derive(InitSpace)]
-pub struct TbrConfigAccount {
+pub struct TbrConfigState {
     /// Program's owner.
     pub owner: Pubkey,
     /// Program's admin. Can be used to update the quotes or appoint a new price updater.
@@ -11,22 +12,17 @@ pub struct TbrConfigAccount {
     /// Intermediate storage for the pending owner. Is used to transfer ownership.
     pub pending_owner: Option<Pubkey>,
 
-    pub quoter_program_address: Pubkey,
     pub fee_recipient: Pubkey,
 
-    //pub token_bridge_address: Pubkey,
-    pub paused_outbound_transfers: bool,
-    /// The fee for the relayer, in μUSD.
-    pub relayer_fee: u64,
     pub evm_transaction_gas: u64,
     pub evm_transaction_size: u64,
 }
 
-impl TbrConfigAccount {
+impl TbrConfigState {
     /// Has any authority:
     /// - Owner
     /// - Admin
-    pub fn is_authorized(&self, key: &Pubkey) -> bool {
+    pub fn is_owner_or_admin(&self, key: &Pubkey) -> bool {
         key == &self.owner || key == &self.admin
     }
 
@@ -39,6 +35,9 @@ impl TbrConfigAccount {
         self.pending_owner.as_ref() == Some(key)
     }
 
-    /// AKA `b"config"`.
-    pub const SEED_PREFIX: &'static [u8; 6] = b"config";
+    /// Value `b"redeemer"`.
+    ///
+    /// It has to stay with that value because that is what the token-bridge is
+    /// looking for.
+    pub const SEED_PREFIX: &'static [u8; 8] = token_bridge::SEED_PREFIX_REDEEMER;
 }
